invoice_be/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-otp==1.2.2
pip==23.2.1
psycopg2==2.9.6
setuptools==67.8.0
sqlparse==0.4.4
tzdata==2023.3
wheel==0.40.0
django-cors-headers==4.3.1
num2words==0.5.13
djangorestframework-simplejwt==5.3.1
pdfkit==1.0.0

